Move into tlf.go
